@@ -25,7 +25,7 @@
 		]
 	},
 	"require": {
-		"php": "^7.3",
+		"php": "^7.4",
 		"cweagans/composer-patches": "^1.7",
 		"filp/whoops": "^2.1.13",
 		"monolog/monolog": "^1.23.0",
