@@ -1554,7 +1554,10 @@ class ilInitialisation
             );
         };
         $c["ui.factory.chart"] = function ($c) {
-            return new ILIAS\UI\Implementation\Component\Chart\Factory($c["ui.factory.progressmeter"]);
+            return new ILIAS\UI\Implementation\Component\Chart\Factory(
+                $c["ui.factory.progressmeter"],
+                $c["ui.factory.bar"]
+            );
         };
         $c["ui.factory.input"] = function ($c) {
             return new ILIAS\UI\Implementation\Component\Input\Factory(
@@ -1610,6 +1613,9 @@ class ilInitialisation
         };
         $c["ui.factory.progressmeter"] = function ($c) {
             return new ILIAS\UI\Implementation\Component\Chart\ProgressMeter\Factory();
+        };
+        $c["ui.factory.bar"] = function ($c) {
+            return new ILIAS\UI\Implementation\Component\Chart\Bar\Factory();
         };
         $c["ui.factory.dropzone.file"] = function ($c) {
             return new ILIAS\UI\Implementation\Component\Dropzone\File\Factory();
